requirements.txt
python-telegram-bot==21.3
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0
gunicorn==21.2.0
telethon==1.36.0
httpx==0.27.2